# Caching of computed resource values from state+changes