This adds a cache for resource values that is evicted by changes to state or plan sync objects.  In theory the eviction is not needed, but there are hints from tests that the current architecture requires it.

Before:
```
goos: linux
goarch: amd64
pkg: github.com/opentofu/opentofu/internal/tofu
cpu: AMD Ryzen 7 2700X Eight-Core Processor         
BenchmarkManyResourceInstances-16              3        13771574147 ns/op
PASS
ok      github.com/opentofu/opentofu/internal/tofu      91.263s
```
After:
```
goos: linux
goarch: amd64
pkg: github.com/opentofu/opentofu/internal/tofu
cpu: AMD Ryzen 7 2700X Eight-Core Processor         
BenchmarkManyResourceInstances-16              5        6983208774 ns/op
PASS
ok      github.com/opentofu/opentofu/internal/tofu      78.373s
```

This is not as dramatic as the first PR, but still has a substantial impact in many workloads.

TODO:
* Experiment with a more granular instance cache suggested by @apparentlymart

## Checklist

<!-- Please check of ALL items in this list for all PRs: -->

- [x] I have read the [contribution guide](https://github.com/opentofu/opentofu/blob/main/CONTRIBUTING.md).
- [x] I have not used an AI coding assistant to create this PR.
- [x] I have written all code in this PR myself OR I have marked all code I have not written myself (including modified code, e.g. copied from other places and then modified) with a comment indicating where it came from.
- [x] I (and other contributors to this PR) have not looked at the Terraform source code while implementing this PR.

### Go checklist

<!-- If your PR contains Go code, please make sure you check off all items on this list: --> 

- [x] I have run golangci-lint on my change and receive no errors relevant to my code.
- [x] I have run existing tests to ensure my code doesn't break anything.
- [x] I have added tests for all relevant use cases of my code, and those tests are passing.
- [x] I have only exported functions, variables and structs that should be used from other packages.
- [x] I have added meaningful comments to all exported functions, variables, and structs.