@cam72cam 2025-10-07
Closing as this does not show a dramatic performance improvement, as well as being obsoleted by the ongoing architecture work.